Made progress on the scraper - see desc.
Scraper currently uses BS4 and then passes to a json. This takes a long time. The YGOProdeck API actually just provides html, no JS, so it could be fine to just use requests.

Passing to JSON also seems to be slow, so maybe it's best to give scraper options for what to extract rather than taking the whole html and parsing it.